APIO-134 move to new structure
