--- modules/apps/web-experience-apio/apio-architect-liferay-portal-impl/src/main/java/com/liferay/apio/architect/liferay/portal/internal/provider/CompanyProvider.java
@@ -1,53 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.apio.architect.liferay.portal.internal.provider;
-
-import com.liferay.apio.architect.provider.Provider;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.Company;
-import com.liferay.portal.kernel.util.Portal;
-
-import javax.servlet.http.HttpServletRequest;
-
-import javax.ws.rs.ServerErrorException;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * Lets resources provide {@code Company} as a parameter to the methods of the
- * different routes builders.
- *
- * @author Alejandro Hernández
- * @author Carlos Sierra Andrés
- * @author Jorge Ferrer
- */
-@Component(immediate = true)
-public class CompanyProvider implements Provider<Company> {
-
-	@Override
-	public Company createContext(HttpServletRequest httpServletRequest) {
-		try {
-			return _portal.getCompany(httpServletRequest);
-		}
-		catch (PortalException pe) {
-			throw new ServerErrorException(500, pe);
-		}
-	}
-
-	@Reference
-	private Portal _portal;
-
-}
